RAG_db/requirements.txt
numpy
sentence-transformers
pdfplumber
ollama
faiss-cpu
chromadb
requests